Working various experience views and searchbar. Heart not showing in favexperiences

+++ getaway/interfaces/src/main/java/ar/edu/itba/getaway/interfaces/persistence/ExperienceDao.java
@@ -15,13 +15,12 @@ public interface ExperienceDao {
 
     // Capaz podría dejarse como el sql
     List<ExperienceModel> listExperiencesByFilter(CategoryModel category, Double max, Long score, CityModel city, Optional<OrderByModel> order, Integer page, Integer page_size);
-
     Long countListByFilter(CategoryModel categoryId, Double max, Long score, CityModel city);
     List<ExperienceModel> listExperiencesByBestRanked(CategoryModel category);
     List<ExperienceModel> listExperiencesFavsByUser(UserModel user, Optional<OrderByModel> order, Integer page, Integer page_size);
     Integer getCountExperiencesFavsByUser(UserModel user);
     List<ExperienceModel> listExperiencesByName(String name, Optional<OrderByModel> order, Integer page, Integer page_size);
-    Integer getCountByName(String name);
+    Long getCountByName(String name);
     boolean hasExperiencesByUser(UserModel user);
     boolean experienceBelongsToUser(UserModel user, ExperienceModel experience);
 }
